+++ src/sessions/video-session-handler.ts
@@ -260,7 +260,8 @@ export default class VideoSessionHandler extends BaseSessionHandler {
       const devices = await getEnumeratedDevices(this.sdk);
       const mediaParams: IMediaRequestOptions = {
         audio: params.audioDeviceId || true,
-        video: params.videoDeviceId || true
+        video: params.videoDeviceId || true,
+        session
       };
 
       if (!devices.videoDevices.length) {
@@ -391,7 +392,9 @@ export default class VideoSessionHandler extends BaseSessionHandler {
       this.log(LogLevels.info, 'Creating new video track', { conversationId: session.conversationId });
 
       // look for a device to use, else use default
-      const track = (await startMedia(this.sdk, { video: params.unmuteDeviceId === undefined ? true : params.unmuteDeviceId })).getVideoTracks()[0];
+      const track = (
+        await startMedia(this.sdk, { video: params.unmuteDeviceId === undefined ? true : params.unmuteDeviceId, session })
+      ).getVideoTracks()[0];
 
       // add track to session
       await this.addReplaceTrackToSession(session, track);
@@ -436,7 +439,9 @@ export default class VideoSessionHandler extends BaseSessionHandler {
         this.log(LogLevels.info, 'No outoing audio to unmute, creating and adding media to session', { sessionId: session.id, conversationId: session.conversationId });
 
         // if params.unmuteDeviceId is `undefined`, use sdk defaults
-        const track = (await startMedia(this.sdk, { audio: params.unmuteDeviceId === undefined ? true : params.unmuteDeviceId })).getAudioTracks()[0];
+        const track = (
+          await startMedia(this.sdk, { audio: params.unmuteDeviceId === undefined ? true : params.unmuteDeviceId, session })
+        ).getAudioTracks()[0];
         await this.addReplaceTrackToSession(session, track);
       }