+++ changelog.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.6.1...HEAD)
+### Fixed
+* make sure output device(s) is available when devices change 
+
+### Changed
+* added session ids to logging around devices
 
 # [v3.6.1](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.6.0...v3.6.1)
 ### Added